Allow using FQDN along with IPv4 and IPv6
This allows a name resolution attempt before checking IP versions.
GetHostByName is already implemented but never been used. When used
against an IP, it will yield the same IP. Therefore no regression
would be implied by this.

When the name resolution returns multiple IPv4 and IPv6 addresses,
GetHostByName only returns the first element of the list which
potentially is going to be an IPv4. This is not very convenient to
have inside GetIp6, but the user can always specify manually the
IPv6 to be used in the config file.