Use getaddrinfo() instaed of gethostbyname()
Like @louisroyer pointed out, gethostbyname() is deprecated and we
should use getaddrinfo() instead.

Note: the default behavior of getaddrinfo() is not the same as
suggested by the man page when hints is set to null. It doesn't
yield IPv6 addresses, and hints.ai_family seems to be set to
AF_INET and not AF_UNSPEC by default.