Add handling of GTP-U Echo Request
As specified in TS 129.281,

```text
Even if there is no path in use, a GTP-U peer shall be prepared to receive an Echo Request at any time and it shall reply with an Echo Response.
```

Also:
```text
4.4.3.2 Echo Response Message
The IP Source Address shall be copied from the IP destination address of the GTP request message to which this GTP-U entity is replying.
The IP Destination Address shall be copied from the IP Source Address of the GTP request message to which this GTP-U entity is replying.
```

With the current implementation, the gnb do not respond to GTP-U Echo Request and log the following message:
```text
[gtp] [error] TEID 0 not found on GTP-U Downlink
```

With this PR, a GTP-U Echo Response is send back to the GTP peer.